% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{Functional interaction analysis of AS events}
\usage{
analyze(query, psi, expr, pathways = NULL, ppi = NULL, restart = 0.7,
  num.folds = 5, num.feats = 100, low.expr = 1, low.var = NULL,
  prop.na = 0.05, prop.extreme = 1, cor.threshold = 0.3)
}
\arguments{
\item{query}{a character vector or data frame containing query genes}

\item{psi}{a SummarizedExperiment object containing PSI of AS events
and an optional data frame describing samples}

\item{expr}{a SummarizedExperiment object or matrix containing gene expression
profiles (FPKM)}

\item{pathways}{a named list of pathway gene sets. If NULL, a combined
list of HALLMARK, KEGG, and REACTOME pathway gene sets will be used.}

\item{ppi}{an \code{igraph} object containing known interactions between
genes. If NULL, an \code{igraph} object containing human gene-gene
interactions will be used.}

\item{restart}{a restart probability}

\item{num.folds}{number of folds for cross-validation}

\item{num.feats}{number of feature nodes to be kept in the final network}

\item{low.expr}{Genes with mean expression below low.expr are excluded. AS
events for corresponding genes are also excluded.}

\item{low.var}{AS events with variance below low.var are excluded. If NULL,
top 10,000 variable events are used for analysis.}

\item{prop.na}{AS events with the higher proportion of missing values than
prop.na are excluded.}

\item{prop.extreme}{AS events with the higher proportion of extreme values
(0 or 1) than prop.extreme are excluded.}

\item{cor.threshold}{a pair of AS event and gene with Spearman's correlation
greather than cor.threshold are connected in a heterogeneous network.}
}
\value{
A matrix containing PSI of AS events
}
\description{
This function analyzes comprehensvie functional interactions of AS events
using Discriminative Random Walk with Restart (DRaWR). It runs a DRaWR on a
heterogeneous network containing genes, AS events, and gene sets
(or pathways). It then performs GSEA on gene sets related to query genes.
}
\details{
This function wraps around the \code{DRaWR} function in the
\code{DRaWR} package with data processing steps.
}
\references{
Blatti, C. et al. (2016). Characterizing gene sets using
discrminative random walks with restart on heterogeneous biological networks
\emph{Bioinformatics}, 32.
}
\keyword{internal}
