% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeFI.R
\name{analyzeFI}
\alias{analyzeFI}
\title{Functional interaction analysis of AS events}
\usage{
analyzeFI(object, query, expr, ppi = NULL, pathways = NULL,
    restart = 0.7, num.folds = 5, num.feats = 100, low.expr = 1,
    low.var = NULL, prop.na = 0.05, prop.extreme = 1,
    cor.threshold = 0.3)
}
\arguments{
\item{object}{Object of class ASpediaFI}

\item{query}{a character vector or a data frame containing query genes}

\item{expr}{a \code{SummarizedExperiment} object or matrix containing gene
expression profiles (FPKM)}

\item{ppi}{an \code{igraph} object containing known interactions between
genes. If NULL, an \code{igraph} object containing human gene-gene
interactions will be used.}

\item{pathways}{a GMT file or a named list of pathway gene sets. If NULL,
a combined list of HALLMARK, KEGG, and REACTOME pathway gene sets will be
used.}

\item{restart}{a restart probability}

\item{num.folds}{the number of folds for cross-validation}

\item{num.feats}{the number of feature nodes to be retained in the final
subnetwork}

\item{low.expr}{Genes with mean expression below low.expr are excluded.
AS events for corresponding genes are also excluded.}

\item{low.var}{AS events with variance below low.var are excluded. If NULL,
top 10,000 variable events are used for analysis.}

\item{prop.na}{AS events with the higher proportion of missing values than
prop.na are excluded.}

\item{prop.extreme}{AS events with the higher proportion of extreme values
(0 or 1) than prop.extreme are excluded.}

\item{cor.threshold}{a pair of AS event and gene with Spearman's correlation
greather than cor.threshold are connected in a heterogeneous network.}
}
\value{
ASpediaFI object with results of functional interaction analysis
}
\description{
Analyze functional interactions of AS events using Discriminative
Random Walk with Restart (DRaWR). It runs a DRaWR on a
heterogeneous network containing genes, AS events, and pathways.
It then performs GSEA on gene sets related to query genes.
}
\examples{
library(limma)
data(GSE114922.fpkm)
data(GSE114922.psi)
design <- cbind(WT = 1, MvsW = colData(GSE114922.psi)$condition == 'MUT')
fit <- lmFit(log2(GSE114922.fpkm + 1), design = design)
fit <- eBayes(fit, trend = TRUE)
tt <- topTable(fit, number = Inf, coef = 'MvsW')
query <- rownames(tt[tt$logFC > 1 & tt$P.Value < 0.1, ])
head(query)
\dontrun{
GSE114922.ASpediaFI <- analyzeFI(
    GSE114922.ASpediaFI, query,
    GSE114922.fpkm
)
}
}
\references{
Blatti, C. et al. (2016). Characterizing
gene sets using discrminative random walks with restart on
heterogeneous biological networks. \emph{Bioinformatics}, 32.
}
