% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoate.R
\name{annotate}
\alias{annotate}
\title{AS event annotation}
\usage{
annotate(ASlist, gtf)
}
\arguments{
\item{ASlist}{a list of AS events}

\item{gtf}{a GRanges object representing GTF}
}
\value{
A list of annotated AS events
}
\description{
This function converts information of AS events into ASpedia format.
}
\keyword{internal}
