% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateASevents.R
\name{annotateASevents}
\alias{annotateASevents}
\title{AS event annotation}
\usage{
annotateASevents(object, gtf.file, num.cores = 1)
}
\arguments{
\item{object}{Object of class ASpediaFI}

\item{gtf.file}{an input GTF file}

\item{num.cores}{the number of cores for parallel processing}
}
\value{
ASpediaFI object with a list of AS event annotations
}
\description{
Detect and annotate AS events from GTF. This function borrows code
from the \code{IVAS} package.
}
\examples{
fi <- new("ASpediaFI")
gtf <- system.file("extdata/GRCh38.subset.gtf", package = "ASpediaFI")
fi <- annotateASevents(fi, gtf.file = gtf, num.cores = 1)
sapply(events(fi), length)
head(events(fi)$SE)
}
\references{
Han, S. et al. (2017). Genome wide
discovery of genetic variants affecting alternative splicing
patterns in human using bioinformatics method. \emph{Genes & Genomics}, 39.
}
