% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateEvent.R
\name{annotateEvent}
\alias{annotateEvent}
\title{AS event annotation}
\usage{
annotateEvent(ASlist)
}
\arguments{
\item{ASlist}{a list of AS events}
}
\value{
A list of annotated AS events
}
\description{
This function converts information of AS events into ASpedia format.
}
\keyword{internal}
