% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/accessors.R
\docType{methods}
\name{samples}
\alias{samples}
\alias{samples,ASpediaFI-method}
\alias{samples<-}
\alias{samples<-,ASpediaFI-method}
\alias{events}
\alias{events,ASpediaFI-method}
\alias{events<-}
\alias{events<-,ASpediaFI-method}
\alias{psi}
\alias{psi,ASpediaFI-method}
\alias{psi<-}
\alias{psi<-,ASpediaFI-method}
\alias{gtf}
\alias{gtf,ASpediaFI-method}
\alias{gtf<-}
\alias{gtf<-,ASpediaFI-method}
\alias{network}
\alias{network,ASpediaFI-method}
\alias{network<-}
\alias{network<-,ASpediaFI-method}
\alias{gene.table}
\alias{gene.table,ASpediaFI-method}
\alias{gene.table<-}
\alias{gene.table<-,ASpediaFI-method}
\alias{as.table}
\alias{as.table,ASpediaFI-method}
\alias{as.table<-}
\alias{as.table<-,ASpediaFI-method}
\alias{pathway.table}
\alias{pathway.table,ASpediaFI-method}
\alias{pathway.table<-}
\alias{pathway.table<-,ASpediaFI-method}
\alias{ASpediaFI-accessor}
\title{ASpediaFI accessors}
\usage{
samples(object, ...)

\S4method{samples}{ASpediaFI}(object)

samples(object) <- value

\S4method{samples}{ASpediaFI}(object) <- value

events(object, ...)

\S4method{events}{ASpediaFI}(object)

events(object) <- value

\S4method{events}{ASpediaFI}(object) <- value

psi(object, ...)

\S4method{psi}{ASpediaFI}(object)

psi(object) <- value

\S4method{psi}{ASpediaFI}(object) <- value

gtf(object, ...)

\S4method{gtf}{ASpediaFI}(object)

gtf(object) <- value

\S4method{gtf}{ASpediaFI}(object) <- value

network(object, ...)

\S4method{network}{ASpediaFI}(object)

network(object) <- value

\S4method{network}{ASpediaFI}(object) <- value

gene.table(object, ...)

\S4method{gene.table}{ASpediaFI}(object)

gene.table(object) <- value

\S4method{gene.table}{ASpediaFI}(object) <- value

as.table(object, ...)

\S4method{as.table}{ASpediaFI}(object)

as.table(object) <- value

\S4method{as.table}{ASpediaFI}(object) <- value

pathway.table(object, ...)

\S4method{pathway.table}{ASpediaFI}(object)

pathway.table(object) <- value

\S4method{pathway.table}{ASpediaFI}(object) <- value
}
\arguments{
\item{object}{an ASpediaFI object}

\item{...}{additional arguments to be passed}

\item{value}{a value to replace. For details, please see help(ASpediaFI).}
}
\value{
Slots of the \code{ASpediaFI} object
}
\description{
ASpediaFI accessors
}
\examples{
fi <- new('ASpediaFI')

data('GSE114922.psi')
psi(fi) <- GSE114922.psi

psi(fi)
}
