% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAnnotationTrack_event.R
\name{createAnnotationTrack_event}
\alias{createAnnotationTrack_event}
\title{Create an event track from AS event}
\usage{
createAnnotationTrack_event(eventGr, type)
}
\arguments{
\item{eventGr}{A GRangesList object of AS event}

\item{type}{AS event type}
}
\value{
an AnnotationTrack object
}
\description{
This function creates an event track from AS event.
}
\details{
This function is borrowed from the \code{maser} package.
}
\references{
Veiga, D. (2019). maser: Mapping Alternative Splicing Events
to pRoteins. R package version 1.2.0. https://github.com/DiogoVeiga/maser
}
\keyword{internal}
