% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAnnotationTrack_transcripts.R
\name{createAnnotationTrack_transcripts}
\alias{createAnnotationTrack_transcripts}
\title{Create transcript tracks from AS event}
\usage{
createAnnotationTrack_transcripts(eventGr, gtf_exons, type)
}
\arguments{
\item{eventGr}{A GRangesList object of AS event}

\item{gtf_exons}{A GRanges object containing exons in gtf}

\item{type}{AS event type}
}
\value{
a list of GeneRegionTrack objects
}
\description{
This function creates transcript tracks from AS event.
}
\details{
This function is borrowed from the \code{maser} package.
}
\references{
Veiga, D. (2019). maser: Mapping Alternative Splicing Events
to pRoteins. R package version 1.2.0. https://github.com/DiogoVeiga/maser
}
\keyword{internal}
