% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createExonTable.R
\name{createExonTable}
\alias{createExonTable}
\title{Create an exon table}
\usage{
createExonTable(gtf_exons, ids)
}
\arguments{
\item{gtf_exons}{A GRanges object containing exons in gtf}

\item{ids}{transcript ids}
}
\value{
a data frame containing an exon table
}
\description{
This function creates an exon table
}
\details{
This function is borrowed from the \code{maser} package.
}
\references{
Veiga, D. (2019). maser: Mapping Alternative Splicing Events
to pRoteins. R package version 1.2.0. https://github.com/DiogoVeiga/maser
}
\keyword{internal}
