% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTxnTrack.R
\name{createTxnTrack}
\alias{createTxnTrack}
\title{Create a transcript track}
\usage{
createTxnTrack(res.df, trackLabel, featureName)
}
\arguments{
\item{res.df}{A data frame containing an exon table}

\item{trackLabel}{Label of a transcript track}

\item{featureName}{Name of feature}
}
\value{
a GeneRegionTrack object
}
\description{
This function creates a transcript track
}
\details{
This function is borrowed from the \code{maser} package.
}
\references{
Veiga, D. (2019). maser: Mapping Alternative Splicing Events
to pRoteins. R package version 1.2.0. https://github.com/DiogoVeiga/maser
}
\keyword{internal}
