% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectEvent.R
\name{detectEvent}
\alias{detectEvent}
\title{Detection of AS events}
\usage{
detectEvent(file, Ncor)
}
\arguments{
\item{file}{Input GTF or GFF3 file.}

\item{Ncor}{The number of cores for parallel processing}
}
\value{
A list of AS events
}
\description{
This function detects AS events from an input GTF.
}
\details{
This function wraps around \code{makeTxDbFromGFF} in the
\code{GenomicFeatures} package and \code{Splicingfinder} in the
\code{IVAS} package.
}
\references{
Han, A. et al. (2017). Genome wide discovery of genetic variants
affecting alternative splicing patterns in human using bioinformatics method.
\emph{Genes & Genomics}, 39.
}
