% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectEventFromTxDb.R
\name{detectEventFromTxDb}
\alias{detectEventFromTxDb}
\title{Detection of AS events from TxDb}
\usage{
detectEventFromTxDb(TxDb, Ncor = 1)
}
\arguments{
\item{TxDb}{A TxDb object}

\item{Ncor}{The number of cores for parallel processing}
}
\value{
A list of AS events
}
\description{
This function detects AS events from a TxDb object.
}
\details{
This function is modified from \code{Splicingfinder} in the
\code{IVAS} package.
}
\references{
Han, A. et al. (2017). Genome wide discovery of genetic variants
affecting alternative splicing patterns in human using bioinformatics method.
\emph{Genes & Genomics}, 39.
}
\keyword{internal}
