% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectRI.R
\name{detectRI}
\alias{detectRI}
\title{Detection of retained introns}
\usage{
detectRI(exon.info, intron.info, alt.intron.info, tx.gene = NULL)
}
\arguments{
\item{exon.info}{A matrix containing exon locus}

\item{intron.info}{A matrix containing intron locus}

\item{alt.intron.info}{A matrix containing flanking introns of
alternative exons}

\item{tx.gene}{A table of transcripts including ID, gene names, etc.}
}
\value{
A matrix containing retained intron events
}
\description{
This function detects retained intron (RI) events.
}
\details{
This function is borrowed from the \code{IVAS} package.
}
\references{
Han, S. et al. (2017). Genome wide discovery of genetic variants
affecting alternative splicing patterns in human using bioinformatics method.
\emph{Genes & Genomics}, 39.
}
\keyword{internal}
