% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectSEandMXE.R
\name{detectSEandMXE}
\alias{detectSEandMXE}
\title{Detection of skipped and mutually exclusive exons}
\usage{
detectSEandMXE(exon.info, intron.info, alt.intron.info)
}
\arguments{
\item{exon.info}{A matrix containing exon locus}

\item{intron.info}{A matrix containing intron locus}

\item{alt.intron.info}{A matrix containing flanking introns of
alternative exons}
}
\value{
A matrix containing SE and MXE events
}
\description{
This function detects skipped (SE) and mutually exclusive exons (MXE) events.
}
\details{
This function is borrowed from the \code{IVAS} package.
}
\references{
Han, A. et al. (2017). Genome wide discovery of genetic variants
affecting alternative splicing patterns in human using bioinformatics method.
\emph{Genes & Genomics}, 39.
}
\keyword{internal}
