% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawr.R
\name{drawr}
\alias{drawr}
\title{Discriminative Random Walk with Restart (DRaWR)}
\usage{
drawr(query.genes, universe, network, restart, num.folds, num.feats)
}
\arguments{
\item{query.genes}{a data frame containing query genes and their weights}

\item{universe}{a data frame containing entire genes}

\item{network}{a data frame containing edges and their weights and types}

\item{restart}{a restart probability}

\item{num.folds}{number of folds for cross-validation}

\item{num.feats}{number of feature nodes to be kept in the final network}
}
\value{
A list of AS events
}
\description{
This function runs the DRaWR on query gene set.
}
\details{
This function is modified from the \code{DRaWR} function in the
\code{DRaWR} package.
}
\references{
Blatti, C. et al. (2016). Characterizing gene sets using
discrminative random walks with restart on heterogeneous biological networks
\emph{Bioinformatics}, 32.
}
\keyword{internal}
