% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportNetwork.R
\name{exportNetwork}
\alias{exportNetwork}
\title{Export network to GML format}
\usage{
exportNetwork(object, node = NULL, file)
}
\arguments{
\item{object}{Object of class ASpediaFI}

\item{node}{the name of pathway. If NULL, the entire subnetwork is exported.}

\item{file}{the file name to export the network}
}
\value{
a GML file containing a subnetwork
}
\description{
Export a subnetwork pertaining to the given pathway to GML format
which can be used in Cytoscape. If no pathway is given, the entire
final subnetwork is exported.
}
\examples{
library(igraph)
fi <- new("ASpediaFI", network = make_empty_graph(n = 0))
exportNetwork(fi, node = NULL, file = "empty.gml")
}
