% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterEvent.R
\name{filterEvent}
\alias{filterEvent}
\title{Detection of AS events from TxDb}
\usage{
filterEvent(ASlist, total.exon.range, total.intron.range, txTable)
}
\arguments{
\item{ASlist}{a list of AS events}

\item{TxDb}{A TxDb object}
}
\value{
A list of AS events
}
\description{
This function detects AS events from a TxDb object.
}
\details{
This function is modified from \code{Splicingfinder} in the
\code{IVAS} package.
}
\references{
Han, S. et al. (2017). Genome wide discovery of genetic variants
affecting alternative splicing patterns in human using bioinformatics method.
\emph{Genes & Genomics}, 39.
}
\keyword{internal}
