% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE114922.fpkm}
\alias{GSE114922.fpkm}
\title{Example gene expression dataset}
\format{An object of class \code{matrix} with 6275 rows and 82 columns.}
\usage{
GSE114922.fpkm
}
\description{
A matrix containing gene expression values. We downloaded RNA-Seq reads of
82 MDS patients from GEO database (GSE114922), aligned with STAR, and obtained
FPKM values using RSEM.
}
\references{
Pellagatti, A. et al. (2018). Impact of spliceosome mutations on
RNA splicing in myelodysplasia: dysregulated genes/pathways and clinical
associations. \emph{Blood}, 132.
}
\keyword{datasets}
