% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE114922.psi}
\alias{GSE114922.psi}
\title{Example dataset containing PSI values}
\format{An object of class \code{SummarizedExperiment} with 10000 rows and 82 columns.}
\usage{
GSE114922.psi
}
\description{
A SummarizedExperiment containing PSI values of 5,000 AS events. We downloaded
RNA-Seq reads of 82 MDS patients from GEO database (GSE114922),
aligned with STAR, and computed PSI values using rMATS. AS events with a lot
of missing values or extreme values, or those on genes with low expression
were filtered out and 5,000 most variable AS events were selected.
}
\references{
Pellagatti, A. et al. (2018). Impact of spliceosome mutations on
RNA splicing in myelodysplasia: dysregulated genes/pathways and clinical
associations. \emph{Blood}, 132.
}
\keyword{datasets}
