% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGRangesListFromEvent.R
\name{makeGRangesListFromEvent}
\alias{makeGRangesListFromEvent}
\title{Make a GRangesList object from an AS event}
\usage{
makeGRangesListFromEvent(event, gtf)
}
\arguments{
\item{event}{AS event ID in ASpedia format}
}
\value{
a GRangesList object
}
\description{
This function makes a GRangesList object from an AS event.
}
\keyword{internal}
