% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapTranscriptsRIevent.R
\name{mapTranscriptsRIevent}
\alias{mapTranscriptsRIevent}
\title{Map transcripts to RI events}
\usage{
mapTranscriptsRIevent(eventGr, gtf_exons)
}
\arguments{
\item{eventGr}{A GRangesList object of AS event}

\item{gtf_exons}{A GRanges object containing exons in gtf}
}
\value{
a list of transcripts
}
\description{
This function maps transcripts to RI events
}
\details{
This function is borrowed from the \code{maser} package.
}
\references{
Veiga, D. (2019). maser: Mapping Alternative Splicing Events
to pRoteins. R package version 1.2.0. https://github.com/DiogoVeiga/maser
}
\keyword{internal}
