% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify.R
\name{quantify}
\alias{quantify}
\title{AS event quantification}
\usage{
quantify(AS.list, sample.info, read.type = "paired", read.length,
  insert.size, min.reads = 3, num.cores = 1)
}
\arguments{
\item{AS.list}{a list of AS events}

\item{sample.info}{a data frame containing names, bam file paths, and
groups (optional) of samples}

\item{read.type}{a type of RNA-seq reads ("single" or "paired")}

\item{read.length}{read length}

\item{insert.size}{insert size}

\item{min.reads}{a minimum number of reads mapped to a given exon}

\item{num.cores}{the number of cores for parallel processing}
}
\value{
A matrix containing PSI of AS events
}
\description{
This function calculates PSI of AS events from input bam files.
}
\details{
This function is modified from the \code{RatioFromReads} function
in the \code{IMAS} package. \code{sample.info} must include two columns
containing names and bam file paths of samples, respectively. The first,
second, and third columns must correspond to names, bam file paths, and groups
of samples.
}
\references{
Han, S. et al. (2017). IMAS: Integrative analysis of
Multi-omics data for Alternative Splicing. R package version 1.8.0.
}
\keyword{internal}
