% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyEvent.R
\name{quantifyEvent}
\alias{quantifyEvent}
\title{AS event quantification}
\usage{
quantifyEvent(ASlist, Total.bamfiles, readsInfo = "paired", readLen,
  insertSize, minr = 3, Ncor = 1)
}
\arguments{
\item{ASlist}{a list of AS events}

\item{Total.bamfiles}{a data frame containing paths and names of bam files}

\item{readsInfo}{a type of RNA-seq reads ("single" or "paired")}

\item{readLen}{read length}

\item{insertSize}{insert size}

\item{minr}{a minimum number of reads mapped to a given exon}

\item{Ncor}{the number of cores for parallel processing}
}
\value{
A list of AS events
}
\description{
This function calculates PSI of AS events from input bam files.
}
\details{
This function is modified from the \code{RatioFromReads} function
in the \code{IMAS} package.
}
\references{
Han, S. and Lee, Y. (2017). IMAS: Integrative analysis of
Multi-omics data for Alternative Splicing. R package version 1.8.0.
}
