% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyPSI.R
\name{quantifyPSI}
\alias{quantifyPSI}
\title{AS event quantification}
\usage{
quantifyPSI(object, read.type = "paired", read.length, insert.size,
  min.reads, num.cores = 1)
}
\arguments{
\item{object}{Object of class ASpediaFI}

\item{read.type}{a type of RNA-seq reads ('single' or 'paired')}

\item{read.length}{read length}

\item{insert.size}{insert size}

\item{min.reads}{a minimum number of reads mapped to a given exon}

\item{num.cores}{the number of cores for parallel processing}
}
\value{
ASpediaFI object with PSI values
}
\description{
Compute PSI values of AS events. This function borrows code
from the \code{IMAS} package.
}
\examples{
bamMT <- system.file("extdata/GSM3167287.subset.bam", package = "ASpediaFI")
bamWT <- system.file("extdata/GSM3167290.subset.bam", package = "ASpediaFI")
GSE114922.ASpediaFI <- ASpediaFI(sample.names = c("GSM3167287",
                                                  "GSM3167290"),
                                 bam.files = c(bamMT, bamWT),
                                 conditions = c("MT", "WT"))
#GSE114922.ASpediaFI <- quantifyPSI(GSE114922.ASpediaFI,
#                                   read.type = "paired",
#                                   read.length = 100, insert.size = 300,
#                                   min.reads = 3, num.cores = 1)
}
\references{
Han, S. et al. (2017). IMAS: Integrative
analysis of Multi-omics data for Alternative Splicing.
R package version 1.8.0.
}
