% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplicingReads.R
\name{SplicingReads}
\alias{SplicingReads}
\title{Counting junction and paired-end reads}
\usage{
SplicingReads(bam.file = NULL, test.exon = NULL, spli.jun = NULL,
  e.ran = NULL, chr = NULL, read.type = "paired", insert.size = 40)
}
\arguments{
\item{bam.file}{a path to bam file}

\item{test.exon}{a data frame containing AS target exon and its neighbors}

\item{spli.jun}{A data frame containing junction information}

\item{e.ran}{A range of parsingr eads from a bam file}

\item{chr}{A chromosome number}

\item{read.type}{a type of RNA-Seq reads}

\item{insert.size}{An insert.sizert size}
}
\value{
A list of counts of junction reads and paired-end reads
}
\description{
This function counts junction and paired-end reads.
}
\details{
This function is borrowed from the \code{IMAS} package.
}
\references{
Han, S. et al. (2017). IMAS: Integrative analysis of Multi-omics
data for Alternative Splicing. R package version 1.8.0.
}
\keyword{internal}
