% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{AS event and pathway visualization}
\usage{
visualize(object, node, zoom = NULL, n = NULL)
}
\arguments{
\item{object}{Object of class ASpediaFI}

\item{node}{the name of AS event or pathway}

\item{zoom}{a logical to determine if genomic coordinates are zoomed
(for AS event visualization)}

\item{n}{the number of genes and AS events to be shown (for pathway
visualization)}
}
\value{
a plot demonstrating AS event or pathway
}
\description{
Visualize AS event or pathway. If an AS event node is given, the
function modified from the \code{plotTranscripts} function in the
\code{maser} package is used to visualize the event. If a
pathway node is given, a subnetwork pertaining to the pathway is
visualized.
}
\examples{
\dontrun{
# Visualize AS event
visualize(GSE114922.ASpediaFI,
    node = as.table(GSE114922.ASpediaFI)$EventID[1],
    zoom = FALSE
)

# Visualize pathway
visualize(GSE114922.ASpediaFI, node = 'HALLMARK_HEME_METABOLISM', n = 10)
}
}
\references{
Veiga, D. (2019). maser: Mapping
Alternative Splicing Events to pRoteins. R package version 1.2.0.
https://github.com/DiogoVeiga/maser
}
