% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeEvent.R
\name{visualizeEvent}
\alias{visualizeEvent}
\title{AS event visualization}
\usage{
visualizeEvent(event, gtf, psi = NULL, zoom = NULL)
}
\arguments{
\item{event}{AS event ID in ASpedia format}

\item{gtf}{a GRanges object representing GTF}

\item{psi}{a SummarizedExperiment object containing PSI of AS events and
groups of samples}

\item{zoom}{a logical to determine if genomic coordinates are zoomed}
}
\value{
a Gviz object
}
\description{
This function visualizes an AS event.
}
\details{
This function is modified from the \code{plotTranscripts} function
in the \code{maser} package.
}
\references{
Veiga, D. (2019). maser: Mapping Alternative Splicing Events
to pRoteins. R package version 1.2.0. https://github.com/DiogoVeiga/maser
}
\keyword{internal}
