% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeNetwork.R
\name{visualizeNetwork}
\alias{visualizeNetwork}
\title{Network visualization}
\usage{
visualizeNetwork(pathway.id, net, n = 10)
}
\arguments{
\item{pathway.id}{name of pathway}

\item{net}{an igraph object containing a heterogeneous network produced
by DRaWR.}

\item{n}{number of genes and AS events to show}
}
\value{
plot of a pathway-specific network
}
\description{
This function visualizes a heterogeneous network pertaining to specific
pathway.
}
\keyword{internal}
