% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplicingReads.R
\name{SplicingReads}
\alias{SplicingReads}
\title{Counting junction and paired-end reads}
\usage{
SplicingReads(bamfile = NULL, test.exon = NULL, spli.jun = NULL,
  e.ran = NULL, SNPchr = NULL, readsinfo = "paired", inse = 40)
}
\arguments{
\item{ASlist}{a list of AS events}

\item{Ncor}{The number of cores for parallel processing}
}
\value{
A list of AS events
}
\description{
This function counts junction and paired-end reads.
}
\details{
This function is borrowed from the \code{IMAS} package.
}
\keyword{internal}
